[bsr][api][features] Refactor database add/removal functions
`legacy_add_feature_to_database()` is only left for _old_ migrations.
It will be removed once those migrations are squashed.

`add_feature_to_database()` is introduced for _future_ migrations.
Since feature flags are now automatically added before deployment,
this function will only be used in the "downgrade" function of
migrations that _remove_ feature flags (to restore the feature flag
when the migration is reverted). It takes a `Feature` object because
we want to control the "isActive" value (which is not possible with
`FeatureToggle`).

`remove_feature_from_database()` is used both for old
migrations (where it takes a FeatureToggle object) and in new
migrations (where it takes a Feature object, for parity with
`add_feature_to_database()`). Its typing has been updated for the
latter usage (because that's the only _current_ usage)