[bsr][api][qa] Unpin pylint and disable a few more checks
We pinned it in 14b78ac448e8e672f7ab3e25 because it crashed but it's
no longer the case.

We now disable a few more checks:

- consider-using-f-string: we do use f-strings where it is pertinent,
but there are a lot of places where it's not. Since we don't want to
clutter the code with "ignore" pragma, we'd rather ignore this check;

- arguments-renamed: it mostly appears in Flask-admin views and in a
few subclasses of our own classes where we provide a name that is more
appropriate than the name used in the method of the parent class. Here
again, adding "ignore" pragma everywhere would be cumbersome. The risk
of having a bug that would have been caught by this check seems
relatively minor.