(BSR)[API] fix: Always use `datetime.utcnow()`, not `now()`, in implementation
This is not a bug fix per se, because the timezone of our production
servers is UTC, which means that `now()` and `utcnow()` are the same.

However, on non-UTC hosts (such as developers' systems), it's not the
same. Considering that datetimes are supposed to be stored in UTC,
using using `now()` here and there (instead of `utcnow()`) could cause
hard to catch test failures or inconsistencies