(PC-13287)[API] feat: Use business unit-venue link date for pricing sort order
Given two bookings B1 and B2 for two venues V1 and V2 of the same
offerer and the following timeline:

1. Booking B1 is marked as used. Its venue V1 is _not_ linked to any
   business unit yet, so B1 is not priced.
2. Booking B2 is marked as used. Its venue V2 is linked to a business
   unit, thus B2 is priced.
3. The offerer links venue V1 to the same business unit as V2.
4. Now B1 can be priced.

With the former version of the code, we had a problem at step 4
because we saw that B2 had a "used" date after B1. So we thought we
could delete it, but that was not possible because B2 could have been
already invoiced. What we do now is that we order bookings on the date
the "business unit-venue link" was created. Here we hence naturally
price B1 after B2, thus leaving B2 untouched.