(BSR)[API] test: Turn warnings into exceptions when running pytest
All warnings have been fixed (or have been "temporarily" silenced with
the `-Wignore` option). To avoid new (and sometimes important*)
warnings popping up and getting unnoticed, let's turn warnings into
exceptions to break the build. Surely we will notice them, now.

We need to use the `-Wignore` flag instead of the `filterwarnings`,
otherwise `-Werror` takes precedence and exceptions are raised for
warnings that we want to ignore.

* Warnings such as "hey, it seems you override a form and disable CSRF
checks: is that intended?"