(BSR)[API] test: Fix bogus test in UbbleWebhookTest
The test was mocking the whole route function... or trying to mock it.
Actually the route was not mocked at all. The test passed nonetheless
because we were really accessing Ubble HTTP API and failed because of
an authentication error (so we did get the error that was expected by
the test).

Now we mock the function that is called by the route, which was the
intention of the test (as far as I can tell, given its name).