(PC-14866)[API] fix: filter out registration date where the user is not eligible
the following happenned : a user creates an account with 14 yearls old but declares a 15 yo age. the user fails educonnect because the age is too young and we override user dateOfBirth with the right dateOfBirth. when the user comes back with 15 years old, in order to give the right amount for deposit, we look for the age at the first registration. If we consider it is 14 yo, we cannot find the right deposit