(PC-15366)[API] fix: Correctly order booking for pricing (again)
The previous change made us order on the business unit-link date first.
This was wrong for the following (rather obvious) case:

1. Venue V1 is linked to business unit BU on 01/01.
2. Venue V2 is then linked to the same BU on 05/01.
3. Booking B2 (on venue V2) is used on 10/01 and is priced.
4. Booking B1 (on venue V2) is used on 15/01.

The previous implementation would declare that bookings of V1 should
be booked before bookings of V2 (because it was linked to its BU
first). That made no sense. The first "component" of the order should
rather be `max(venue_link_date, used_date)` (in fact, something a
little more complex but that's the idea).