(BSR)[API] fix: Fix looping in `price_bookings()`
We used to base loops on the booking id. That was wrong. For example,
suppose that all bookings were sorted like this for pricing:

   1, 4, 2, 3

With a batch size of 2, the first batch would be [1, 4]. The last id
being for, we would add a "id > 4" filter for the second batch. The
second batch would thus be empty.

Now we add a filter that uses the "composite key" with which bookings
are sorted: it's based on the "dateUsed", the date of the pricing
point link, the booking id, etc.