(PC-16633)[API] perf: Optimize `_generate_cashflows()` pricings query
The business unit-based SQL query in `_generate_cashflows()` was fast
because we filtered on `pricing.businessUnitId`, which has an index.

With reimbursement points, we did not use any index, so we JOINed a
lot of pricings with a lot of bookings. This took several seconds
for each reimbursement point to process.

This commit adds a `pricing.venueId` column with an index, on which
we can now quickly filter.