(PC-17328)[API] fix: Force pricing of unused-then-used bookings
Bookings that have been used, then unused and then used again were not
priced. This was a bug, it has been fixed in 312bbb7551399ba80b8b5393.

The pricing of some of these bookings is now blocked because dependent
bookings have now been reimbursed (we cannot delete their pricings).
We have to force the pricing of these bookings.