(PC-17170)[API] feat: no error if email update already done
Do not raise a UserDoesNotExist when changing a user's email if the
change has already been done: the update request exists inside the
user's email update history and has been validated.

It seems that some users click more than one time on the validation link
which generates some errors that should not be. The email update has
been validated: no need to do it again and no need to raise an error.