(PC-18410)[ADMIN] FlaskAdmin: deactivate CSRF protection...
...because Flask-WTF, which has been added recently, already handles it.

Flask-WTF adds some defaults (config and jinja template filters) which
do not seem to work very well the default behaviour of FlaskAdmin's
forms and CSRF protection: two csrf tokens were created for each form.
Each form is then considered as invalid. And no update/patch can by
applied to any object because of this messed up CSRF protection.

The idead is to deactivate the CSRF protection from FlaskAdmin and let
Flask-WTF handle it globally. Which seems to work well: only one CSRF
token is generated and the protection is enabled.