(BSR)[API] chore: Temporarily disable new mypy "truthy-function" error
Starting from version 0.990, mypy reports a new mypy "truthy-function"
error. A lot of occurrences of this code (probably all of them,
considering that "real" errors should already be detected by a pylint
checker) are false positives that happen on code that use methods
decorated with SQLAlchemy's `hybrid_property`. Example:

    class User:
        @hybrid_property
        def has_admin_role(self) -> bool:
            ...

        @has_admin_role.expression  # type: ignore [no-redef]
        def has_admin_role(cls) -> bool:
            ...

        def func(self):
            if self.has_admin_role:
                ...

mypy thinks that `self.has_admin_role` is a method (while it's in fact
a property), and hence wrongfully reports an error:

    Function "Callable[..., Any]" could always be true in boolean context [truthy-function]