(PC-18610)[API] feat: Check pricing integrity before generating cashflows
We should not modify the amount of a booking after it's been priced
(at least, not without re-pricing the booking). Unfortunately we did,
once.

To detect that, a check is added before generating the cashflow: we
look at the "offerer revenue" pricing line, which should be the same
as the booking amount. If it's not, we log an error and refuse to
generate the cashflow for the concerned reimbursement point.