(PC-16737)[API] perf: Optimize generation of wallets CSV file
The generation of this CSV took between 10 and 15 minutes on
production. This is because we were calling the `get_wallet_balance()`
SQL function for each user (thus selecting the deposit and then
selecting related bookings for each user): that takes time.

And on staging the process was even OOMKilled, most probably because
it loaded all rows in memory (and our staging pods have less memory).

The calculation of the wallet balance is now done by a single query
that can join everything for all users. The SQL query is long but the
whole generation takes around 2 minutes on production. And we don't
keep all rows in memory anymore (making it work on staging).