(PC-18614)[API] feat: Require price in API Stocks (under feature flag)
Until now, the price was optional in "API Stocks". In December, we'll
toggle a feature flag to make it mandatory.

We want a feature flag so that we can easily toggle it off if problems
arise. Because of this, we cannot provide a true API schema through
Swagger that says whether the price is required or optional. Instead,
the schema still says that the price is optional and we trigger a
special validator if the feature flag is on. The textual description
of the schema has been updated to reflect about this change (and users
have been notified).

Once we're sure that all providers have switched, we'll remove the
feature flag and use a default validator.