(BSR)[API] test: Remove unnecessary calls to `freeze_time()`
We have used `freeze_time()` because we wanted to check that a
specific datetime was stored. This caused an issue when we froze
the time more than 2 years in the past: we would create a user
with a 2-year old deposit that would be expired when the test
tried to create a booking, which caused an `insufficientFunds` in the
`check_booking` PostgreSQL function.

Instead of looking for a specific datetime, we can merely check that
we have a non-null value. It's good enough in most cases.