(PC-19610)[API] feat: Add safety belt to avoid duplicate invoices
With the previous version of the code, if `flask generate_invoices`
was executed twice simultaneously, it could be possible to have
duplicate invoices for the same cashflow. From an accounting point of
view, this should not be possible and should be avoided.

The new version of the code acquires an xact lock related to the
reimbursement point when we start to process it and double checks the
absence of related invoice.