(BSR)[API] feat: Translate pydantic error message when None is not an allowed value
If the client sends `None` and pydantic returns a `none.not_allowed`
error, it's likely that the field is required. But perhaps there are
places where we don't allow None but allow the empty string? I am not
sure why we would do that. Anyway, this is why the message is not "Ce
champ est obligatoire", but rather "Ce champ ne peut pas être nul".