(BSR)[API] feat: Simplify access to `User.roles`
The column is not nullable. By making it the empty list by default
when the model is instantiated, we can simplify code. (No, providing a
`default` value on the SQLAlchemy field does not set the value soon
enough.)

Having to initialize the `roles` attribute in Flask-Admin code is
unfortunate, but that code will soon go away. We can live with that
for a few weeks.