(BSR)[API] chore: Remove unused offerer/venue search features
The home page of the pro portal used to let admin users search for
offerers and venues by keywords. It looked for these keywords in the
name, "public name", city, address, SIREN and SIRET. It was very slow
and barely usable. This page has been removed in dd4ff5477fc2cb8f0481.

As such, we can remove that search code, as well as the underlying
PostgreSQL indexes. Search is now performed from the back-office and
uses other indexes.