(PC-22790)[API] feat: Introduce finance events
This commit introduces the notion of "finance event", which inserts
itself between a booking (and, in the future, anything that can be
priced) and a pricing. A finance event is thus created:

- when a booking is marked as used;
- when a used booking is cancelled or un-unused (back to "confirmed");
- when a booking has been reimbursed and then cancelled.

Once a finance event is created, it's ready to be priced by a new
`price_finance_events` cronjob. The rest of the finance workflow
(creation of cashflow from pricings, etc.), models and their statuses
do not change.

A `PRICE_FINANCE_EVENTS` feature flag is created to control the switch
from `price_bookings` cronjob to `price_finance_events`:

1. PRICE_BOOKINGS is enabled: `price_bookings` runs,
   `price_finance_events` does not run.

2. Populate the `finance_event` table, so that all pricings have a
   corresponding event.

3. Disable `PRICE_BOOKINGS`. From now on, no pricings will be directly
   created.

4. Finish populating the `finance_event` table (same as step 2, for
   pricings that have been created between step 2 and step 3).

5. Enable `PRICE_FINANCE_EVENTS`.