Revert "(PC-21704)[API] feat: Limit number of offerers in "/offerer/names" for admin users"
This reverts commit 2c02a45a11a640d6cc6fb7006a2ab3ae23d6cc23.

The commit caused only 50 offerers to be shown in the dropdown of the
home page (`/accueil`"). That works well when no offerer is selected.
However, the same page is used to show a particular offerer
(`accueil?structure=17`): when that offerer is not amongst the 50
returned ones, the dropdown cannot show it as the selected option and
hence falls back to selecting the first option. This is problematic
because the name of the offerer is thus not displayed anywhere.