(BSR)[API] style: singleton-comparison: Use `Model.column.is_()` and `isnot()`
Part 2 of enabling `singleton-comparison` pylint warning.

We now enforce the use of `Model.column.is_(None)` in SQLAlchemy
filters instead of `Model.column == None`. While the latter works, it
yields the `singleton-comparison` warning in Pylint. Also, it could
easily be (mis-)rewritten as `Model.column is None`, which always
evaluates to True. Ditto for `isnot()` instead of `!=`.

When the column is a boolean, this commits also uses the shorter form
`Model.column` in filters (instead of `Model.column.is_(True)`), as
well as `sa.not_(Model.column)`.